feat: additional refactoring of comment-related widgets
