+++ lib/src/features/comment/presentation/widgets/widgets.dart
@@ -3,7 +3,9 @@ export 'comment_bottom_sheet/comment_action_bottom_sheet.dart';
 export 'comment_bottom_sheet/comment_comment_action_bottom_sheet.dart';
 export 'comment_bottom_sheet/general_comment_action_bottom_sheet.dart';
 export 'comment_card/comment_card.dart';
-export 'comment_card/comment_card_actions.dart';
-export 'comment_card/comment_card_header.dart';
+export 'comment_card/comment_card_button_actions.dart';
+export 'comment_card/comment_card_header/comment_card_header.dart';
 export 'comment_card/comment_content.dart';
 export 'comment_card/comment_depth_indicator.dart';
+export 'comment_card/additional_comment_card.dart';
+export 'comment_card/comment_card_background.dart';
