+++ lib/src/features/search/presentation/pages/search_page.dart
@@ -715,11 +715,7 @@ class _SearchPageState extends State<SearchPage> with AutomaticKeepAliveClientMi
                           10,
                         ),
                       ),
-                      CommentListEntry(
-                        comment: comment,
-                        onVoteAction: (int commentId, int voteType) => context.read<SearchBloc>().add(VoteCommentEvent(commentId: commentId, score: voteType)),
-                        onSaveAction: (int commentId, bool save) => context.read<SearchBloc>().add(SaveCommentEvent(commentId: commentId, save: save)),
-                      ),
+                      CommentListEntry(comment: comment),
                     ],
                   );
                 }
